🚧 api request

+++ flutter_module/api/weather_api/lib/src/model/rsp_get_geocoding_location.dart
@@ -23,7 +23,7 @@ class Location extends Equatable {
   final String? name;
   final double? latitude;
   final double? longitude;
-  final int? elevation;
+  final double? elevation;
   final String? featureCode;
   final String? countryCode;
   final int? admin1Id;
