+++ flutter_module/api/weather_api/lib/src/model/rsp_get_weather_forecast.dart
@@ -11,7 +11,7 @@ class Weather {
   int? utcOffsetSeconds;
   String? timezone;
   String? timezoneAbbreviation;
-  int? elevation;
+  double? elevation;
   CurrentWeather? currentWeather;
 
   Weather(
